# public/openapi.yaml
openapi: 3.0.3
info:
  title: Finanzas App API
  description: |
    API REST de la aplicación de finanzas personales (Next.js App Router + Prisma + PostgreSQL).

    ## Autenticación
    La mayoría de los endpoints requieren un **access token JWT** en el header
    `Authorization: Bearer <token>`. El token se obtiene en `/auth/login` o `/auth/register`.
    El **refresh token** se entrega como cookie `HttpOnly` (`refresh_token`) y se usa en `/auth/refresh`.

    ## Convenciones
    - Montos monetarios: número decimal (la API los serializa a veces como string desde PostgreSQL).
    - Fechas: ISO 8601 (`YYYY-MM-DDTHH:mm:ss.sssZ`).
    - Errores de validación → `422` con `{ error, details }`.
  version: 1.0.0
servers:
  - url: http://localhost:3001/api
    description: Desarrollo local
...
security:
  - bearerAuth: []
paths:
...
  /msi:
    get:
      tags: [MSI]
      summary: Listar compras a meses
      responses:
        '200': { description: Lista, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/MSICredit' } } } } }
    post:
      tags: [MSI]
      summary: Registrar compra a meses
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [accountId, description, totalAmount, months, startDate]
              properties:
                accountId:    { type: string }
                description:  { type: string, maxLength: 200 }
                vendor:       { type: string }
                totalAmount:  { type: number }
                months:       { type: integer, minimum: 1, maximum: 60 }
                interestRate: { type: number, nullable: true, description: '0–5 (decimal). Vacío = sin intereses (MSI)' }
                startDate:    { type: string, format: date-time }
      responses:
        '201': { description: Creada, content: { application/json: { schema: { $ref: '#/components/schemas/MSICredit' } } } }
        '422': { $ref: '#/components/responses/ValidationError' }
  /msi/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [MSI]
      summary: Detalle de un MSI
      responses: { '200': { description: MSI }, '404': { $ref: '#/components/responses/NotFound' } }
    patch:
      tags: [MSI]
      summary: Actualizar MSI (ej. paidMonths, status)
      responses: { '200': { description: Actualizado }, '404': { $ref: '#/components/responses/NotFound' } }
    delete:
      tags: [MSI]
      summary: Cancelar MSI
      responses: { '200': { description: Cancelado }, '404': { $ref: '#/components/responses/NotFound' } }

  # ─────────────────── RECURRING INCOME ───────────────────
...
  /recurring-income/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    patch:
      tags: [Recurring Income]
      summary: Actualizar ingreso recurrente
      responses: { '200': { description: Actualizado }, '404': { $ref: '#/components/responses/NotFound' } }
    delete:
      tags: [Recurring Income]
      summary: Desactivar ingreso recurrente
      responses: { '200': { description: Desactivado }, '404': { $ref: '#/components/responses/NotFound' } }
...
  /movimientos:
    get:
      tags: [Análisis]
      summary: Movimientos de una categoría (detalle del slice)
      description: |
        Lista los gastos de una sola categoría en el rango, para el acordeón/detalle
        de Análisis. Mismo filtro que `/analisis/gastos`, por lo que la suma de filas
        coincide con el slice del donut.
      parameters:
        - { name: categoria_id, in: query, schema: { type: string }, description: 'Omitido / "null" / "none" → bucket "Sin categoría" (categoryId null)' }
        - { name: rango, in: query, schema: { type: string, enum: [semana, mes, semestre, anio], default: mes } }
      responses:
        '200':
          description: Movimientos de la categoría
          content:
            application/json:
              schema:
                type: object
                properties:
                  categoriaId: { type: string, nullable: true }
                  rango: { type: string, enum: [semana, mes, semestre, anio] }
                  from:  { type: string, format: date-time }
                  to:    { type: string, format: date-time }
                  total: { type: number }
                  count: { type: integer }
                  movimientos:
                    type: array
                    description: Ordenado por fecha desc
                    items:
                      type: object
                      properties:
                        id:       { type: string }
                        fecha:    { type: string, format: date-time }
                        concepto: { type: string }
                        monto:    { type: number }
                        notes:    { type: string, nullable: true }
                        account:
                          type: object
                          properties:
                            id:    { type: string }
                            name:  { type: string }
                            color: { type: string }
                            type:  { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ───────────────────────── SPLIT ────────────────────────
...
  /notifications/push:
    post:
      tags: [Notifications]
      summary: Suscribirse a push
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [endpoint, p256dh, auth]
              properties:
                endpoint:  { type: string, format: uri }
                p256dh:    { type: string }
                auth:      { type: string }
                userAgent: { type: string }
      responses: { '200': { description: Suscripción guardada } }
    delete:
      tags: [Notifications]
      summary: Cancelar suscripción push
      responses: { '200': { description: Suscripción eliminada } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Access token JWT obtenido en /auth/login o /auth/register'
  responses:
    Unauthorized:
      description: Token ausente, inválido o expirado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: Datos inválidos
      content:
        application/json:
          schema:
            type: object
            properties:
              error:   { type: string, example: Datos inválidos }
              details: { type: object, additionalProperties: true }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
        code:  { type: string }
...
    MSICredit:
      type: object
      properties:
        id:              { type: string }
        accountId:       { type: string }
        description:     { type: string }
        vendor:          { type: string, nullable: true }
        totalAmount:     { type: number }
        months:          { type: integer }
        monthlyAmount:   { type: number }
        interestRate:    { type: number, nullable: true }
        startDate:       { type: string, format: date-time }
        endDate:         { type: string, format: date-time }
        paidMonths:      { type: integer }
        monthsRemaining: { type: integer }
        progressPercent: { type: integer }
        pendingTotal:    { type: number }
        status:          { type: string, enum: [ACTIVE, COMPLETED, CANCELLED] }
...